inventario-nft-juego/frontend/requirements.txt
pygame==2.1.2
requests==2.26.0
python-dotenv==0.19.0 